% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullclean.R
\docType{data}
\name{fullclean}
\alias{fullclean}
\title{fullclean: <Título descriptivo de los datos>}
\format{
Un data frame con <número> filas y <número> columnas:
\describe{
\item{\if{html}{\out{<ID>}}}{Número de identificación del sujeto}
\item{\if{html}{\out{<EDAD>}}}{Edad del sujeto. Tipo: numérico}
\item{\if{html}{\out{<TIEMPODOC>}}}{Tiempo de docencia, en años. Tipo: <numérico/factor/etc.>}
\item{\if{html}{\out{<SEXO>}}}{<descripción de la columna 4>. Tipo: <numérico/factor/etc.>}
\item{\if{html}{\out{<SEXO>}}}{<descripción de la columna 4>. Tipo: <numérico/factor/etc.>}... agrega todas las columnas que tengas

}
}
\source{
<Origen de los datos. Ej: "Datos simulados para propósitos de ejemplo",
o "Estudio realizado en X población", o "Tomado de la publicación Y">
}
\usage{
data(fullclean)
}
\description{
<Descripción extensa del dataset. Por ejemplo:>
Este conjunto de datos contiene los resultados de un análisis factorial
utilizado como ejemplo en las funciones del paquete \code{facomplex}.
Incluye las cargas factoriales de <número> ítems en <número> factores.
}
\examples{
data(fullclean)
head(fullclean)
summary(fullclean)

}
\keyword{datasets}
