% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KC.R
\name{print.KC}
\alias{print.KC}
\title{Print method for KC objects}
\usage{
\method{print}{KC}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"KC"}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object \code{x} (called for side effects of printing).
}
\description{
Print method for KC objects
}
